perf(java): store the projection in the width its own capacity rule admits
ProjectionCapacity already refuses any graph whose vertex count or entry count
does not fit an array index, so every projection this library can hold is one
whose identifiers fit thirty-two bits. The CSR stored them in sixty-four, which
wasted half of the topology and half of the memory traffic that reaches it.

The storage is now int; the public accessors still speak long, because the width
of a vertex identifier is an implementation detail and callers should not have to
track it. The endpoint buffers a merged build fills narrow the same way, so a
build peaks at half of what it did. The bulk export widens on the way out, which
also makes it defensive without the constructor copying the whole graph.

Measured on twenty million vertices and sixty million edges under -Xmx4g: the
projection settles at 560 MB instead of 1.12 GB, peaks at 1.12 GB instead of
2.24 GB, and builds in 3.7 s instead of 18.5 s. The build is five times faster
from a two-fold cut, because the placement pass is bound by memory bandwidth and
the per-vertex sort now compares half as many bytes. A bounded expansion answers
in 3 us instead of 6, and a hundred-thousand-node component in 23 ms instead
of 137.

Directive: keep the storage width and ProjectionCapacity.MAX_VERTICES in step. If
a future projection is to hold more vertices than an array index addresses, the
storage has to leave the heap first, not merely widen.

+++ maven-projects/reader/src/main/java/org/apache/graphar/reader/ProjectionCapacity.java
@@ -29,8 +29,9 @@ package org.apache.graphar.reader;
  * a statement of what was attempted and what the limit is, checked before any array is allocated.
  *
  * <p>The numbers are exact, not estimates: a CSR of {@code v} vertices and {@code e} adjacency
- * entries occupies {@code 8 * (v + 1)} bytes of offsets and {@code 8 * e} bytes of destinations. An
- * undirected projection stores each edge twice, so its entry count is twice its edge count.
+ * entries occupies {@code 4 * (v + 1)} bytes of offsets and {@code 4 * e} bytes of destinations,
+ * because a projection this class admits is one whose identifiers fit an array index. An undirected
+ * projection stores each edge twice, so its entry count is twice its edge count.
  */
 public final class ProjectionCapacity {
     /** The largest number of vertices a heap CSR can address. */
@@ -39,7 +40,7 @@ public final class ProjectionCapacity {
     /** The largest number of adjacency entries a heap CSR can address. */
     public static final long MAX_ENTRIES = Integer.MAX_VALUE - 8L;
 
-    private static final long BYTES_PER_SLOT = Long.BYTES;
+    private static final long BYTES_PER_SLOT = Integer.BYTES;
 
     private ProjectionCapacity() {}
